Move all media selector filtering/sorting algorithm to MediaSelectorFilterSortAlgorithm
