+++ app/shared/app-data/src/commonMain/kotlin/domain/mediasource/MediaListFilter.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2024 OpenAni and contributors.
+ * Copyright (C) 2024-2025 OpenAni and contributors.
  *
  * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
  * Use of this source code is governed by the GNU AGPLv3 license, which can be found at the following link.
@@ -25,7 +25,7 @@ import me.him188.ani.datasources.api.topic.EpisodeRange
  *
  * 使用 [BasicMediaListFilter] 可简化类型声明.
  *
- * @see DefaultRssMediaSourceEngine
+ * @see me.him188.ani.app.domain.mediasource.rss.DefaultRssMediaSourceEngine
  */
 fun interface MediaListFilter<in Ctx : MediaListFilterContext> {
     /**
